limes: add additional aggregations of commitment numbers
For the domain-level and cloud-level UIs, they will serve to generate
overviews, before drilling down into the commitments of specific
projects.

For the project-level UI, they will remove the need for the UI to query
for the full list of commitments only to mark resources that have
pending or planned commitments at all.