liquid: add ResourceInfo.Attributes
As explained in the code comment, this came out of a redundancy that I
noticed when writing liquid-ironic.

These attributes will initially not be visible on the Limes v1 API, but
I plan to use them as an input for the TranslationRule when translating
v2 subcapacities and subresources back into their v1 equivalents.