liquid: adopt type Option as a replacement for pointers
This replaces all places where pointer types are used to denote
optionality throughout all LIQUID types. Places in which pointers are
used only for interior mutability are left as-is.

`Rate.PerAZ[].Usage` is a bit of a weird case: Previously, we used nil to
denote absence, so this becomes an `Option[]` type. But the payload type
is still `*big.Int` because the API for big.Int only allows bigints to
be passed by reference. A validation is added to forbid the value
`Some(nil)`. At the same time, I'm adding a validation to forbid `None`
for rates with `HasUsage = true`.