limesresources: add UUID field to commitments
It is a historical mistake of mine that we exposed numerical IDs on
commitments in the first place. We should have UUIDs here to avoid
exposing the number of commitments to non-privileged users, and as
defense in depth against enumeration attacks.

Because commitments of some sort will start appearing in LIQUID soon,
I want to right this wrong before it seeps into the LIQUID API. LIQUID
will only ever refer to commitments by UUID, and the Limes v2 API will
also move to only using UUIDs. The v1 API will show both identifiers to
aid in the migration.