liquid: use omitempty on newly added field
When working on the prod outage, I noticed that my limesctl choked on
ServiceInfo returned from prod, because prod already had this new field,
but I had not updated go-api-declarations in limesctl.

This is not nice, and we should make a slight bit of effort to be a
slight bit more backwards-compatible.