liquid: add ResourceName.IsValid(), RateName.IsValid()
Because the "service/resource" notation for resources is
well-established, and the "service/resource/az" notation for
AZ resources is establishing itself, I would like to explicitly forbid
slashes from occurring in service types and resource names.

If we're doing a validation at all, I feel like we can be slightly more
restrictive than just "no slashes", though, so this commit establishes a
rule that resource names and rate names must basically look like C
identifiers, except that dashes and dots are also allowed. This matches
all existing resource names in production.

For rate names, the situation is a bit weirder, because we have two
competing precedents:

- For rates reporting usage (currently only Ceph), rate names look like
  resource names, e.g. `attachment_size` or `recipients`.
- For rates with limits (currently only Swift in QA), rate names look
  like CADF actions, e.g.  `account/container:write`.

The second format here has the significant disadvantage of allowing (and
even requiring) slashes in the name, which is specifically what I want
to forbid. I choose to go with the first format here, so rates following
the second format will have to be renamed. This ought to be fine since
the existing Limes integration in sapcc/openstack-rate-limit-middleware
is not actually used at this point. If it is to be used, we can expose
the CADF action name as a separate field on the Limes API.